+++ packages/jupyterlab-fonts/src/stylist.ts
@@ -1,17 +1,18 @@
-import { Notebook, NotebookPanel, NotebookPanel as panel } from '@jupyterlab/notebook';
+import { Notebook, NotebookPanel } from '@jupyterlab/notebook';
 import { Signal } from '@lumino/signaling';
 import * as JSS from 'jss';
 import jssPresetDefault from 'jss-preset-default';
 
 import * as SCHEMA from './schema';
 
-import { ROOT, IFontFaceOptions, DOM } from '.';
+import { ROOT, IFontFaceOptions, DOM, PACKAGE_NAME } from '.';
 
 export class Stylist {
   fonts = new Map<string, IFontFaceOptions>();
 
   private _globalStyles: HTMLStyleElement;
-  private _notebookStyles = new Map<panel, HTMLStyleElement>();
+  private _notebookStyles = new Map<NotebookPanel, HTMLStyleElement>();
+  private _transientNotebookStyles = new Map<NotebookPanel, SCHEMA.ISettings>();
   private _jss = JSS.create(jssPresetDefault());
   private _fontCache = new Map<string, SCHEMA.IFontFacePrimitive[]>();
   private _cacheUpdated = new Signal<this, void>(this);
@@ -76,13 +77,44 @@ export class Stylist {
     return Array.from(this._notebookStyles.keys());
   }
 
-  stylesheet(meta: SCHEMA.ISettings, panel?: NotebookPanel, clear = false) {
+  setTransientNotebookStyle(
+    panel: NotebookPanel,
+    style: SCHEMA.ISettings | null
+  ): void {
+    if (style == null) {
+      this._transientNotebookStyles.delete(panel);
+    } else {
+      this._transientNotebookStyles.set(panel, style);
+    }
+    const meta = panel.model?.metadata.get(PACKAGE_NAME) as SCHEMA.ISettings;
+    this.stylesheet(meta, panel);
+  }
+
+  getTransientNotebookStyle(panel: NotebookPanel): SCHEMA.ISettings | null {
+    return this._transientNotebookStyles.get(panel) || null;
+  }
+
+  stylesheet(meta: SCHEMA.ISettings | null, panel?: NotebookPanel, clear = false) {
     let sheet = panel ? this._notebookStyles.get(panel) : this._globalStyles;
 
-    let style = panel ? this._nbMetaToStyle(meta, panel) : this._settingsToStyle(meta);
+    let style: SCHEMA.IStyles | null = null;
+    let jss: JSS.StyleSheet | null = null;
+    let css: string = '';
 
-    let jss = this._jss.createStyleSheet(style as any);
-    let css = jss.toString();
+    if (meta) {
+      style = panel ? this._nbMetaToStyle(meta, panel) : this._settingsToStyle(meta);
+      jss = this._jss.createStyleSheet(style as any);
+      css = jss.toString();
+    }
+
+    if (panel) {
+      let transientMeta = this.getTransientNotebookStyle(panel);
+      if (transientMeta) {
+        style = this._nbMetaToStyle(transientMeta, panel);
+        jss = this._jss.createStyleSheet(style as any);
+        css = `${css}\n${jss.toString()}`;
+      }
+    }
 
     if (sheet && sheet.textContent !== css) {
       sheet.textContent = css;
@@ -95,8 +127,8 @@ export class Stylist {
     meta: SCHEMA.ISettings,
     notebook: NotebookPanel
   ): SCHEMA.IStyles {
-    const id = notebook.id;
     let jss: any = { '@font-face': [], '@global': {}, '@import': [] };
+    const id = notebook.id;
     let idStyles: any = (jss['@global'][`.${DOM.notebookPanel}[id='${id}']`] = {});
 
     if (meta.fonts) {